requirements.txt
streamlit
tensorflow
Pillow
matplotlib
seaborn